requirements.txt
typer
rich
paramiko
reportlab
python-docx
scapy
selenium
